update: refactor name of tab to randomgame
