+++ app/src/main/java/com/example/project1/ui/randomgame/RandomGameFragment.java
@@ -1,4 +1,4 @@
-package com.example.project1.ui.notifications;
+package com.example.project1.ui.randomgame;
 
 import android.os.Bundle;
 import android.view.LayoutInflater;
@@ -14,17 +14,17 @@ import androidx.lifecycle.ViewModelProviders;
 
 import com.example.project1.R;
 
-public class NotificationsFragment extends Fragment {
+public class RandomGameFragment extends Fragment {
 
-    private NotificationsViewModel notificationsViewModel;
+    private RandomGameViewModel randomGameViewModel;
 
     public View onCreateView(@NonNull LayoutInflater inflater,
                              ViewGroup container, Bundle savedInstanceState) {
-        notificationsViewModel =
-                ViewModelProviders.of(this).get(NotificationsViewModel.class);
-        View root = inflater.inflate(R.layout.fragment_notifications, container, false);
-        final TextView textView = root.findViewById(R.id.text_notifications);
-        notificationsViewModel.getText().observe(getViewLifecycleOwner(), new Observer<String>() {
+        randomGameViewModel =
+                ViewModelProviders.of(this).get(RandomGameViewModel.class);
+        View root = inflater.inflate(R.layout.fragment_randomgame, container, false);
+        final TextView textView = root.findViewById(R.id.text_randomgame);
+        randomGameViewModel.getText().observe(getViewLifecycleOwner(), new Observer<String>() {
             @Override
             public void onChanged(@Nullable String s) {
                 textView.setText(s);
