+++ app/src/main/res/navigation/mobile_navigation.xml
@@ -18,8 +18,8 @@
         tools:layout="@layout/fragment_dashboard" />
 
     <fragment
-        android:id="@+id/navigation_notifications"
-        android:name="com.example.project1.ui.notifications.NotificationsFragment"
-        android:label="@string/title_notifications"
-        tools:layout="@layout/fragment_notifications" />
+        android:id="@+id/navigation_randomgame"
+        android:name="com.example.project1.ui.randomgame.RandomGameFragment"
+        android:label="@string/title_randomgame"
+        tools:layout="@layout/fragment_randomgame" />
 </navigation>
