+++ app/src/main/res/values/strings.xml
@@ -2,5 +2,5 @@
     <string name="app_name">Project1</string>
     <string name="title_phonebook">PhoneBook</string>
     <string name="title_dashboard">Pictures</string>
-    <string name="title_notifications">Notifications</string>
+    <string name="title_randomgame">Random Game</string>
 </resources>
